+++ crates/cli/tests/reference/targets-target-web-experimental-reset-state-function.js
@@ -6,7 +6,6 @@ export function __wbg_reset_state () {
     const wasmInstance = new WebAssembly.Instance(wasmModule, __wbg_get_imports());
     wasm = wasmInstance.exports;
     wasm.__wbindgen_start();
-
 }
 
 /**
@@ -44,19 +43,10 @@ function __wbg_get_imports() {
 
 let __wbg_instance_id = 0;
 
-let cachedInt32ArrayMemory0 = null;
-function getInt32ArrayMemory0() {
-    if (cachedInt32ArrayMemory0 === null || cachedInt32ArrayMemory0.byteLength === 0) {
-        cachedInt32ArrayMemory0 = new Int32Array(wasm.memory.buffer);
-    }
-    return cachedInt32ArrayMemory0;
-}
-
 let wasmModule, wasm;
 function __wbg_finalize_init(instance, module) {
     wasm = instance.exports;
     wasmModule = module;
-    cachedInt32ArrayMemory0 = null;
     wasm.__wbindgen_start();
     return wasm;
 }