+++ CHANGELOG.md
@@ -15,6 +15,15 @@
   exporting it as a public export.
   [#5081](https://github.com/wasm-bindgen/wasm-bindgen/pull/5081)
 
+* Reinitialization is no longer automatically applied when using `panic=unwind`
+  and `--experimental-reset-state-function`, instead it is triggered by any
+  use of the `handler::schedule_reinit()` function under `panic=unwind`,
+  which is supported from within the `on_abort` handler for reinit workflows.
+  Renamed `handler::reinit()` to `handler::schedule_reinit()` and removed
+  the `set_on_reinit()` handler. The `__instance_terminated` address
+  is now always a simple boolean (`0` = live, `1` = terminated).
+  [#5083](https://github.com/wasm-bindgen/wasm-bindgen/pull/5083)
+
 ### Fixed
 
 * Fixed two CLI issues affecting WASM modules built by rustc 1.94+. First,